+++ src/nix/lock.rs
@@ -1,19 +1,19 @@
-use std::process::{Command, Stdio};
+use std::process::Stdio;
 
 use anyhow::{bail, Result};
+use nix_rs::command::NixCmd;
 
 use crate::nix::util::print_shell_command;
 
 use super::url::FlakeUrl;
 
-pub fn nix_flake_lock_check(url: &FlakeUrl) -> Result<()> {
-    let args = ["flake", "lock", "--no-update-lock-file", &url.0];
-    print_shell_command("nix", args.into_iter());
-    let status = Command::new("nix")
-        .args(args)
-        .stdin(Stdio::null())
-        .spawn()?
-        .wait()?;
+/// Make sure that the `flake.lock` file is in sync.
+pub async fn nix_flake_lock_check(url: &FlakeUrl) -> Result<()> {
+    let nix = NixCmd::default();
+    let mut cmd = nix.command();
+    cmd.args(&["flake", "lock", "--no-update-lock-file", &url.0]);
+    print_shell_command(&cmd);
+    let status = cmd.stdin(Stdio::null()).spawn()?.wait().await?;
     if status.success() {
         Ok(())
     } else {