+++ src/nix/util.rs
@@ -1,22 +1,11 @@
 use colored::Colorize;
+use tokio::process::Command;
 
 /// Print a shell command that the user can run as-is.
-pub fn print_shell_command<'a>(program: &str, args: impl Iterator<Item = &'a str>) {
-    let color = |s: &str| s.blue().bold();
-    eprintln!(
-        "> {} {}",
-        color(program),
-        color(
-            &args
-                // If the argument contains a special character, it must be
-                // quoted. We currently check only for "?", though.
-                .map(|x| if x.contains('?') {
-                    format!("\"{}\"", x)
-                } else {
-                    x.to_string()
-                })
-                .collect::<Vec<_>>()
-                .join(" ")
-        )
-    );
+///
+/// FIXME: Well, this doesn't produce a copyable CLI. Just a debug dump of
+/// `Command`.
+pub fn print_shell_command(cmd: &Command) {
+    let cmdline = format!("{:?}", cmd);
+    eprintln!("> {}", cmdline.blue().bold());
 }